% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Morphology_Profile.R
\docType{data}
\name{Cell_Morphology_Profile}
\alias{Cell_Morphology_Profile}
\title{{Single cell morphological profile}}
\format{An object of class \code{data.frame} with 163 rows and 812 columns.}
\usage{
Cell_Morphology_Profile
}
\description{
{
The single cell morphological profiles of human bone osteosarcoma cells (U-2 OS cells) in response to the drugs or small molecule compounds from the library of network-based cellular signatures (LINCS) database (http://lincscloud.org). 
This sample data set is subset of original data including the profile of 9515 drugs or small molecule compounds.
}
}
\keyword{datasets}

