% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Morphology_Profile_Complete.R
\docType{data}
\name{Cell_Morphology_Profile_Complete}
\alias{Cell_Morphology_Profile_Complete}
\title{{Single cell morphological profile}}
\format{An object of class \code{data.frame} with 9515 rows and 812 columns.}
\usage{
Cell_Morphology_Profile_Complete
}
\description{
{
The single cell morphological profiles of human bone osteosarcoma cells (U-2 OS cells) in response to the drugs or small molecule compounds from the library of network-based cellular signatures (LINCS) database (http://lincscloud.org). 
}
}
\keyword{datasets}

