% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{cellMorphologyEnrichmentAnalysis}
\alias{cellMorphologyEnrichmentAnalysis}
\title{cell morphology enrichment analysis}
\usage{
cellMorphologyEnrichmentAnalysis(number_of_features)
}
\arguments{
\item{number_of_features}{The number of top cell morphological features, ranked based on the Strength Centrality Score (SCS) for enrichment analysis (e.g. 20).
This parameter specifies the number of first top cell morphological features which are used for enrichment sets of landmark genes.}
}
\value{
1. A data frame including 812 cell morphological features (row) and associated land mark genes with each feature (column); 2. A data frame including profiles of cell morphological features of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column cell morphological features); 3. A data frame including gene expression profiles of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column gene symbol).
}
\description{
We use a stepwise variable selection approach, including combination of least absolute shrinkage and selection operator (LASSO) with cross-validation to tune parameter for cell morphology enrichment analysis. We consider all transcriptomic profiles in the reference repository as inputs of LASSO to select a subset of landmark genes (v) that best describe an indicated profile of cell morphology feature.
}
\examples{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
data(Query_Transcriptomic_Profile)
cellMorphologyEnrichmentAnalysis(20)
}
\author{
{Isar Nassiri, Matthew McCall}
}
