% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{crosstabulation}
\alias{crosstabulation}
\title{{Cross-tabulation of landmark genes, and single-cell morphological features}}
\usage{
crosstabulation()
}
\description{
{We present the results of cell morphology enrichment analysis as cross-tabulation of landmark genes, and single-cell morphological features including the direction of effects (up or down regulation).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
TOP=10
input="BRD-K37798499"
crosstabulation() 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

