% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{GRN}
\alias{GRN}
\title{{Gene regulatory network of cell morphological phenotypes}}
\usage{
GRN(number_of_features, support, confidence)
}
\arguments{
\item{number_of_features:}{An integer, number of cell morphological phenotype; support: An integer, is the proportion of cell morphological features which are associated with an indicated gene set; confidence: An integer, the threshold of confidence.}
}
\description{
{We use the mining association model to detect the associations between the landmark genes, and inference of gene regulatory network of cell morphological phenotypes. .}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
Mapping("BRD-K37798499")
GRN(10, 0.2, 0.9) 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

