% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{Mapping}
\alias{Mapping}
\title{{Mapping query transcriptomic profile against the reference repository}}
\usage{
Mapping(input)
}
\arguments{
\item{input:}{A character, name of a drug or small compound molecule (e.g. "BRD-K37798499").}
}
\description{
{We map the query profile (Q) of fold change of gene expression versus reference repository of transcriptomic data (T) to detect similarities among these profiles (s).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
Mapping("BRD-K37798499") 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

