% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{mappingQueryTranscriptomic}
\alias{mappingQueryTranscriptomic}
\title{Mapping query transcriptomic profile against the reference repository}
\usage{
mappingQueryTranscriptomic(Query_Transcriptomic_Profile, Transcriptomic_Profile,
  Cell_Morphology_Profile)
}
\arguments{
\item{Query_Transcriptomic_Profile}{A data frame including expression level of 978 land mark genes in response to treatment with an indicated drugs/small compound molecule (row drug/small molecule compound ID, and column land mark gene symbols)}

\item{Transcriptomic_Profile}{A data frame including expression level of 978 land mark genes in response to treatment with 162 drugs/small compound molecules (row drug/small molecule compound ID, and column land mark gene symbols)}

\item{Cell_Morphology_Profile}{A data frame including profiles of 812 cell morphological features of 162 drugs/small compound molecules (row drug/small molecule compound ID, and column cell morphological features)}
}
\value{
1. A data frame including profiles of cell morphological features of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column cell morphological features); 3. A data frame including gene expression profiles of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column gene symbol).
}
\description{
{We map the query profile (Q) of fold change of gene expression versus reference repository of transcriptomic data (T) to detect similarities among these profiles (s).
First, we convert the query and backend repository of transcription profiles to the Boolean expression and replace up-regulated values with one and down regulated
values with zero. We provide a confusion matrix for the intersection of the query with each of the transcriptomic profiles in the reference repository.
Next, we use the confusion matrices to score the similarity between the query and reference transcription profiles based on the Matthew correlation value.
}
}
\examples{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
data(Query_Transcriptomic_Profile)
mappingQueryTranscriptomic(Query_Transcriptomic_Profile, Transcriptomic_Profile, Cell_Morphology_Profile)
}
\author{
{Isar Nassiri, Matthew McCall}
}
