% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{{CellMorphologyEnrichmentAnalysis}}
\alias{{CellMorphologyEnrichmentAnalysis}}
\alias{CellMorphologyEnrichmentAnalysis}
\title{{Cell morphology enrichment analysis}}
\usage{
CellMorphologyEnrichmentAnalysis(number_of_features)
}
\arguments{
\item{number_of_features}{A Number}
}
\description{
{We use a stepwise variable selection approach, including combination of least absolute shrinkage and selection operator (LASSO) with cross-validation to tune parameter for cell morphology enrichment analysis. We consider all transcriptomic profiles in the reference repository as inputs of LASSO to select a subset of landmark genes (v) that best describe an indicated profile of cell morphology feature. }
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
Mapping("BRD-K37798499")
CellMorphologyEnrichmentAnalysis(20) 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}
