% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{{crossTabulation}}
\alias{{crossTabulation}}
\alias{crossTabulation}
\title{{Cross-tabulation of landmark genes, and single-cell morphological features}}
\usage{
crossTabulation(TOP, input)
}
\arguments{
\item{TOP}{A Number}

\item{input}{A Name}
}
\description{
{We present the results of cell morphology enrichment analysis as cross-tabulation of landmark genes, and single-cell morphological features including the direction of effects (up or down regulation).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
Mapping("BRD-K37798499")
CellMorphologyEnrichmentAnalysis(20) 
crossTabulation(10, "BRD-K37798499") 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}
