% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{{GRN}}
\alias{{GRN}}
\alias{GRN}
\title{{Gene regulatory network of cell morphological phenotypes}}
\usage{
GRN(number_of_features, support, confidence, input)
}
\arguments{
\item{number_of_features}{A Number}

\item{support}{A Number}

\item{confidence}{A Number}

\item{input}{A character}
}
\description{
{We use the mining association model to detect the associations between the landmark genes, and inference of gene regulatory network of cell morphological phenotypes. .}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
GRN(10, 0.1, 0.6, "BRD-K37798499") 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}
