% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{{RankCellMorphologicalFeatures}}
\alias{{RankCellMorphologicalFeatures}}
\alias{RankCellMorphologicalFeatures}
\title{{Ranking of cell morphological phenotypes based on the Strength Centrality Score (SCS).}}
\usage{
RankCellMorphologicalFeatures(TOP, number_of_features, input)
}
\arguments{
\item{TOP}{A Number}

\item{number_of_features}{A Number}

\item{input}{A character}
}
\description{
{We consider the connectivity and centrality of cell morphological features in Cosine similarity network of image-based cell morphological profile to rank them based on the Strength Centrality Score (SCS).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
RankCellMorphologicalFeatures(20, 20, "BRD-K37798499") 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}
