% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query_Transcriptomic_Profile.R
\docType{data}
\name{Query_Transcriptomic_Profile}
\alias{Query_Transcriptomic_Profile}
\title{{Query Transcriptomic Profile}}
\format{An object of class \code{matrix} with 978 rows and 1 columns.}
\usage{
Query_Transcriptomic_Profile
}
\description{
{
This file contains a vector of landmark gene expression level in response to the treatment with a therapeutic agent.
The sample input includes transcriptomic profile of U-2 OS cells in response to the treatment with ETOPOSIDE (BRD-K37798499). 
ETOPOSIDE is considered an effective agent in combined chemotherapy of osteosarcoma patients.
}
}
\references{
{
1. Gill, J., Ahluwalia, M.K., Geller, D. and Gorlick, R. (2013) New targets and approaches in osteosarcoma. Pharmacology & Therapeutics, 137, 89-99.
}
}
\keyword{datasets}
