% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{Ranking_Cell_Morphological_features}
\alias{Ranking_Cell_Morphological_features}
\title{{Ranking of cell morphological phenotypes based on the Strength Centrality Score (SCS).}}
\usage{
Ranking_Cell_Morphological_features()
}
\description{
{We consider the connectivity and centrality of cell morphological features in Cosine similarity network of image-based cell morphological profile to rank them based on the Strength Centrality Score (SCS).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
TOP=20
Ranking_Cell_Morphological_features() 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

