% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcriptomic_Profile.R
\docType{data}
\name{Transcriptomic_Profile}
\alias{Transcriptomic_Profile}
\title{{Transcriptomic Profiles of U-2 OS cells in response to the perturbations with drugs and small molecules}}
\format{An object of class \code{data.frame} with 163 rows and 978 columns.}
\usage{
Transcriptomic_Profile
}
\description{
{
The transcriptomic profiles of human bone osteosarcoma cells (U-2 OS cells) in response to the 9515 drugs or small molecule compounds from the library of network-based cellular signatures (LINCS) database (http://lincscloud.org).
This sample data set is subset of original data including the profile of 9515 drugs or small molecule compounds.
}
}
\keyword{datasets}

