% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{crosstabulation}
\alias{crosstabulation}
\title{{Cross-tabulation of landmark genes, and single-cell morphological features}}
\usage{
crosstabulation(TOP, input)
}
\arguments{
\item{input:}{An integer, numeber of top cell morphological phenotypes in the ranked list, input: A character, name of a drug or small molecule.}
}
\description{
{We present the results of cell morphology enrichment analysis as cross-tabulation of landmark genes, and single-cell morphological features including the direction of effects (up or down regulation).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
Mapping("BRD-K37798499")
Cell_Morphology_Enrichment_Analysis(20) 
crosstabulation(10, "BRD-K37798499") 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

