% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{crossTabulation}
\alias{crossTabulation}
\title{{Cross-tabulation of landmark genes, and single-cell morphological features}}
\usage{
crossTabulation(TOP)
}
\arguments{
\item{TOP}{We use the top cell morphological features, ranked based on the Strength Centrality Score (SCS) for enrichment analysis.
This parameter specifies the number of first top cell morphological features which are used for enrichment sets of landmark genes.}
}
\value{
1. A crosstab matrix including 812 cell morphological features (row), associated land mark genes with each feature (column), and gene expression level vaules as its elements; 2. A data frame including profiles of cell morphological features of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column cell morphological features); 3. A data frame including gene expression profiles of similar drugs/small compound molecules with query (row drug/small molecule compound ID, and column gene symbol).
}
\description{
We present the results of cell morphology enrichment analysis as cross-tabulation of landmark genes, and single-cell morphological features including the direction of effects (up or down regulation).
}
\examples{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
data(Query_Transcriptomic_Profile)
crossTabulation(20) 
}
\author{
{Isar Nassiri, Matthew McCall}
}
