% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{Mapping}
\alias{Mapping}
\title{{Mapping query transcriptomic profile against the reference repository}}
\usage{
Mapping()
}
\description{
{We map the query profile (Q) of fold change of gene expression versus reference repository of transcriptomic data (T) to detect similarities among these profiles (s).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
input="BRD-K37798499"
Mapping() 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}

