% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMEA.R
\name{{crossTabulation}}
\alias{{crossTabulation}}
\alias{crossTabulation}
\title{{Cross-tabulation of landmark genes, and single-cell morphological features}}
\usage{
crossTabulation(TOP)
}
\arguments{
\item{TOP}{A Number}
}
\description{
{We present the results of cell morphology enrichment analysis as cross-tabulation of landmark genes, and single-cell morphological features including the direction of effects (up or down regulation).}
}
\examples{
{
data(Transcriptomic_Profile)
data(Cell_Morphology_Profile)
data(Query_Transcriptomic_Profile)
crossTabulation(20) 
}
}
\author{
{Isar Nassiri, Matthew McCall}
}
